Create a new component to display the license info modal for reusability
